Separated Pagination logic into util
